Added new fields agar, gcs and fhr

+++ database/migrations/2021_07_14_140219_create_covids_table.php
@@ -22,7 +22,7 @@ class CreateCovidsTable extends Migration
             $table->integer('age');
             $table->string('unit');
             $table->string('sex');
-            $table->integer('contact_of_caller');
+            $table->bigInteger('contact_of_caller');
             $table->string('location');
             $table->string('complaints');
             $table->string('vaccine');
